Crate sudoku

Sudoku

A Sudoku solver that uses a non-recursive backtracking (depth-first search) algorithm.

Typical usage

Ingest an unsolved puzzle into a Grid using one of the available methods or trait implementations of Grid. Depending on the input type, 0 or '0' should be used to represent unfilled cells.

Call Grid::solve to find a solution.

Display the result using the Display trait or access the values using the Grid::as_slice method.

Examples

Ingesting a string and pretty-printing the solution:

use sudoku::Grid;

let puzzle: Grid = " \
    003020600 \
    900305001 \
    001806400 \
    008102900 \
    700000008 \
    006708200 \
    002609500 \
    800203009 \
    005010300"
    .parse()
    .unwrap();
let solution = puzzle.solve().expect("No solution found");
println!("Solution:\n{}", solution);

The above code will print:

483|921|657
967|345|821
251|876|493
---+---+---
548|132|976
729|564|138
136|798|245
---+---+---
372|689|514
814|253|769
695|417|382

Ingesting a u8 array and reading the first digit of the solution:

use sudoku::Grid;

let puzzle: Grid = Grid::from_array(&[
    0, 0, 3, 0, 2, 0, 6, 0, 0, 9, 0, 0, 3, 0, 5, 0, 0, 1, 0, 0, 1, 8, 0, 6, 4, 0, 0, 0, 0,
    8, 1, 0, 2, 9, 0, 0, 7, 0, 0, 0, 0, 0, 0, 0, 8, 0, 0, 6, 7, 0, 8, 2, 0, 0, 0, 0, 2, 6,
    0, 9, 5, 0, 0, 8, 0, 0, 2, 0, 3, 0, 0, 9, 0, 0, 5, 0, 1, 0, 3, 0, 0,
]);
let solution = puzzle.solve().expect("No solution found");
assert_eq!(4_u8, solution.as_slice()[0]);

Structs

Grid

A container to hold a Sudoku grid.