Module sudoku::grid

Contain the main utility for Sudoku

Structs

BackTracePositionTracker

An iterator that can move forward or backwward. It is used to track the position in the back trace method

CellCoordinate

Represent a coordinate on a crate::grid::Sudoku grid. It is a number between 0 and [GAME_SIZE] - 1 (which is 8).

CellPosition

Represent a crate::cell::Cell position on a crate::grid::Sudoku

Column

Column iterator of a crate::grid::Sudoku. It iterate over the second coordinate.

Row

Row iterator of a crate::grid::Sudoku. It iterate over the first coordinate.

Square

Square iterator of a crate::grid::Sudoku.

Sudoku

Represent the grid with the cell and the current game state

Enums

Direction

Represent either a forward or backward direction

SolveError

Error returned by Sudoku::solve_back_trace

VerificationError

Error of a Sudoku configuration, the first cell position is the place of the stronger cell

VerificationResult

Contain a verification sucess but either for a complete configuration (or sub configuration) or incomplete.

Traits

SudokuIter

This trait cannot be implemented outside the crate as it depend on a the Sealed which is private